Updated file reading and MediaType assignment logic
Updated the URL string in `Program.cs` to read from a local file path instead of a web URL. Enhanced the condition in `FileReader.cs` for calling `GetStreamFromLocalFileAsync(path)` to include cases where the URI scheme is a file scheme, allowing local files to be read in addition to web resources. In `HttpArchiveReader.cs`, added a condition to check and set the `MediaType` of `parametricContent` based on the "Content-Type" header if it's not initially provided in the request's postData.