`Added PostmanCollectionReader class with incomplete method`
Added a new file `PostmanCollectionReader.cs` to the project, which includes a new class `PostmanCollectionReader` under the namespace `Hmb.Biriba.SpecFormats.PostmanV2_1_0`. This class has a private readonly field `_jsonSerialization` of type `JsonSerialization` and a constructor that initializes this field. The class also includes a public asynchronous method `GetParametricRequestsAsync` that currently returns an empty sequence, indicating an incomplete implementation. Several namespaces have been imported at the beginning of the file.