Refactor StringArrayJsonConverter and add new classes
Refactored `StringArrayJsonConverter` in `PostmanCollectionConverters.cs` to include two new methods for handling string arrays. Added `CharSeparatedStringArrayJsonConverter` and `DotSeparatedStringArrayJsonConverter` classes that inherit from `StringArrayJsonConverter` and provide specific functionality for character-separated and dot-separated strings. Introduced an abstract `Separator` property in `CharSeparatedStringArrayJsonConverter` for defining the separation character. Added `DotSeparatedStringArrayJsonConverter` and `SlashSeparatedStringArrayJsonConverter` classes that override the `Separator` property to provide '.' and '/' separators. Updated `JsonConverter` attribute for `Host` and `Path` properties in `PostmanCollectionModels.cs` to use the new classes for processing dot-separated and slash-separated strings.