Add `ExtensionData` property to `PostmanCollectionModels.cs`
A new property `ExtensionData` has been added to the `PostmanCollectionModels.cs` file in the `Hmb.Biriba.SpecFormats.PostmanV2_1_0` namespace. This property is of type `Dictionary<string, JsonElement>?` and is decorated with the `JsonExtensionData` attribute. This attribute indicates that this property should be used to store any additional JSON properties that are not already covered by other class properties when the class is deserialized. This is particularly useful for handling dynamic JSON where some data elements may not be known in advance.