`Update file paths and conditions in PostmanCollectionReader`
Modified the `postmanFile` path in `Program.cs` from a URL to a local file path. In `PostmanCollectionReader.cs`, updated the condition to check if `item.Request.Body` is not null. The condition now checks for `item.Request.Body.Raw` and `item.Request.Body.Mode` as `raw`. Also, added conditions to check for `item.Request.Body.File` and `item.Request.Body.GraphQl` with respective modes. The `parametricContent.Content` for these conditions is commented out.