Subject: Refactor MediaType setting in PostmanCollectionReader
Body:
Updated the `PostmanCollectionReader` class to enhance flexibility in setting the `MediaType` of `parametricContent`. A new check has been added to assign `Content-Type` from `parametricRequest` to `parametricContent.MediaType` if present. The `MediaType` is no longer hardcoded in `BodyMode` checks (`raw`, `urlencoded`, `formdata`, `file`, `graphql`), instead, it checks for `contentType` value and falls back to the previous hardcoded value if not present. This allows `MediaType` to be set from the `Content-Type` header or default to a value based on `BodyMode`.